Second Draft of Presentation
A bunch of changes based on initial feedback and suggestions. This includes
some structural and organizational changes as well as a bunch of images that
have been added to the presentation to liven it up a little bit and keep people
interested (hopefully). Also added a few diagrams and the graffle file I used
to create all of them (graffle file contains _all_ of the diagrams, each on a
separate layer).